pyproject.toml
[project]
name = "vibe-blender"
version = "0.1.0"
description = "A fully automated, headless Text-to-3D agentic tool that generates Blender models via ReAct self-correction"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "openai>=1.0",
    "httpx>=0.25",
    "pillow>=10.0",
    "imageio>=2.31",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[project.scripts]
vibe-blender = "vibe_blender.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vibe_blender"]
sources = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
